# resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.1.0",
        "title": "Azure DocumentDB Connector"
    },
    "host": "<host>",
    "paths": {
        "/docs/{id}": {
            "get": {
                "operationId": "Get Document",
                "description": "Gets a document by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "x-ms-summary": "Document ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved document",
                        "schema": {
                            "$ref": "#/definitions/document"
                        },
                        "headers": {
                            "x-ms-doc-id": {
                                "description": "ID of the document",
                                "x-ms-summary": "ID",
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Update Document",
                "description": "Update a document",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "x-ms-summary": "Document ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved document",
                        "schema": {
                            "$ref": "#/definitions/document"
                        },
                        "headers": {
                            "x-ms-doc-id": {
                                "description": "ID of the document",
                                "x-ms-summary": "ID",
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "Delete Document",
                "description": "Deletes a document",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "x-ms-summary": "Document ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/docs": {
            "post": {
                "operationId": "Create Document",
                "description": "Create a document",
                "parameters": [
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created document",
                        "schema": {
                            "$ref": "#/definitions/document"
                        },
                        "headers": {
                            "x-ms-doc-id": {
                                "description": "ID of the document",
                                "x-ms-summary": "ID",
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "operationId": "Upsert Document",
                "description": "Upsert a document",
                "parameters": [
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Upserted document",
                        "schema": {
                            "$ref": "#/definitions/document"
                        },
                        "headers": {
                            "x-ms-doc-id": {
                                "description": "ID of the document",
                                "x-ms-summary": "ID",
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/query/docs": {
            "post": {
                "operationId": "Query Documents",
                "description": "DocumentDB SQL Query",
                "parameters": [
                    {
                        "name": "x-ms-account",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Account",
                        "description": "Account name",
                        "required": true
                    },
                    {
                        "name": "x-ms-masterkey",
                        "in": "header",
                        "type": "string",
                        "description": "Primary or Secondary Key",
                        "x-ms-summary": "Key",
                        "required": false
                    },
                    {
                        "name": "x-ms-dbs",
                        "in": "header",
                        "type": "string",
                        "description": "Database name",
                        "x-ms-summary": "Database",
                        "required": true
                    },
                    {
                        "name": "x-ms-colls",
                        "in": "header",
                        "type": "string",
                        "x-ms-summary": "Collection",
                        "description": "Collections name",
                        "required": true
                    },
                    {
                        "name": "body",
                        "description": "A valid documentDb body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "query": {
                                    "type": "string",
                                    "description": "DocumentDB SQL Query"
                                },
                                "parameters": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "x-ms-summary": "Object Array",
                                        "description": "Name/Value objects to substitue out for query",
                                        "x-ms-visibility": "advanced"
                                    }
                                }
                            },
                            "required": [
                                "query"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Query response",
                        "schema": {
                            "$ref": "#/definitions/queryResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "document": {
            "type": "object",
            "additionalProperties": {
                "type": [
                    "string",
                    "boolean",
                    "number",
                    "object",
                    "array"
                ]
            }
        },
        "queryResponse": {
            "type": "object",
            "properties": {
                "_rid": {
                    "type": "string",
                    "x-ms-summary": "Resource ID",
                    "description": "The resource id for the collection used within the query.",
                    "x-ms-visibility": "advanced"
                },
                "_count": {
                    "type": "integer",
                    "description": "Number of documents returned",
                    "x-ms-summary": "Count"
                },
                "Documents": {
                    "type": "array",
                    "description": "Document set",
                    "items": {
                        "$ref": "#/definitions/document"
                    }
                }
            },
            "required": [
                "_rid",
                "_count",
                "Documents"
            ]
        }
    }
}